flatten event + query handlers
