+++ server/tests/test_apps/history/conftest.py
@@ -7,7 +7,6 @@ from sqlalchemy import text
 
 from the_history_atlas.apps.history import HistoryApp
 from the_history_atlas.apps.history.database import Database
-from the_history_atlas.apps.history.query_handler import QueryHandler
 from the_history_atlas.apps.history.trie import Trie
 
 
@@ -17,12 +16,12 @@ def mock_db():
 
 
 @pytest.fixture
-def readmodel_app(engine, config):
+def history_app(engine, config):
     return HistoryApp(database_client=engine, config_app=config)
 
 
 @pytest.fixture
-def readmodel_db(engine):
+def history_db(engine):
     source_trie = Trie()
     entity_trie = Trie()
     database = Database(
@@ -44,23 +43,7 @@ def source_author():
 
 
 @pytest.fixture
-def query_handler(engine):
-    source_trie = Trie()
-    entity_trie = Trie()
-    database = Database(
-        database_client=engine, source_trie=source_trie, entity_trie=entity_trie
-    )
-    source_trie.build(entity_tuples=database.get_all_source_titles_and_authors())
-    entity_trie.build(entity_tuples=database.get_all_entity_names())
-    return QueryHandler(
-        database_instance=database,
-        source_trie=source_trie,
-        entity_trie=entity_trie,
-    )
-
-
-@pytest.fixture
-def cleanup_tag(readmodel_db) -> Generator[Callable[[UUID], None], None, None]:
+def cleanup_tag(history_db) -> Generator[Callable[[UUID], None], None, None]:
     tags_to_cleanup: set[UUID] = set()
 
     def _cleanup(tag_id: UUID) -> None:
@@ -69,7 +52,7 @@ def cleanup_tag(readmodel_db) -> Generator[Callable[[UUID], None], None, None]:
     yield _cleanup
     if not tags_to_cleanup:
         return  # don't error on an empty tuple
-    with readmodel_db.Session() as session:
+    with history_db.Session() as session:
         # cleanup
         session.execute(
             text(
